+++ src/Loaders/LoadingSpinner.spec.tsx
@@ -1,9 +1,7 @@
 import * as faker from "faker";
 import React from "react";
-import { cleanup, render } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import { LoadingSpinner } from "./LoadingSpinner";
-
-afterEach(cleanup);
 
 test("it should have `role='progress'`", () => {
   const { getByRole } = render(<LoadingSpinner />);